refactor: localStorage 관리 함수 모아둔 곳으로부터 불러오게 해라
