+++ src/app/asset/validation/checkProductName.js
@@ -1,6 +1,6 @@
 import { createCheckEveryFunction, createValidFunction, isRequired } from './common.js';
 import { ERROR_MSG } from '../constants/index.js';
-import { isUniqueProductName } from '../../localStorage/inventory.js';
+import { isUniqueProductName } from '../../localStorage/index.js';
 
 const checkProductName = createCheckEveryFunction([
     createValidFunction(isRequired, ERROR_MSG.requireProductName),
