+++ src/app/controller/Purchase.js
@@ -1,10 +1,15 @@
 import PurchaseTap from '../view/PurchaseTap.js';
 import { checkInputAmount } from '../asset/validation/index.js';
-import { getInputAmount, setInputAmount } from '../localStorage/inputAmount.js';
-import { getProducts, setProducts } from '../localStorage/inventory.js';
+import {
+    getInputAmount,
+    setInputAmount,
+    getProducts,
+    setProducts,
+    getCoins,
+    setCoins,
+} from '../localStorage/index.js';
 import { ERROR_MSG } from '../asset/constants/index.js';
 import { distributeCoinGreedily, calcCoinAmount } from '../asset/util/index.js';
-import { getCoins, setCoins } from '../localStorage/coin.js';
 
 export default class Purchase {
     constructor($skeleton) {
